+++ NegocioAlumnos/NegocioAlumnos.cs
@@ -82,7 +82,7 @@ namespace Negocio
                 throw;
             }
         }
-        public static int registrarEstado(string estado, string fecha, float dni)
+        public static int registrarEstado(string estado, string fecha, int dni)
             {
             
             try
